requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
openpyxl>=3.1.0
python-dateutil>=2.8.0

# ML libraries (required for model loading)
xgboost>=2.0.0
lightgbm>=4.0.0

# Additional utilities
joblib>=1.3.0

# Daily update system
schedule>=1.2.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Web scraping (for daily updates)
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Multi-modal model dependencies
tensorflow>=2.13.0  # Use tensorflow-gpu>=2.13.0 for GPU support
tensorflow-gpu>=2.13.0  # Optional: for GPU acceleration
Pillow>=10.0.0  # Image processing
catboost>=1.2.0  # CatBoost ensemble model
tqdm>=4.66.0  # Progress bars
